LookAndFeel not possible for now.
If all else fails we will have to create a new class with custom parameters, pass an instance of that new
class to a 2nd class that derives from LookAndFeel but is never exposed to Lua, and when a method from the
C++ class is called by a component to paint something we will call methods from the passed in instance by
name (or some other faster way). This might affect performance but for now i have no other ideas on how
to do this.