I'll be moving the debugger to the Lua Editor window, so that we can re-use the code editor components in it. No need to write some new stuff if that's already there.
Also the Linux crash when debugging paint() callbacks seems to be something i can't resolve and Jules@juce forum
does not have a clue about. So i need to test on windows to see if similar stuff happens.

Also the clidebugger.lua should be tested as it provides more feature we could use (breakpoints in code, instead
of calling dbg() and some other additional stuff we could use)